案例审核

+++ smile_admin/src/main/java/org/flightythought/smile/admin/database/repository/RecoverCaseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecoverCaseRepository extends JpaRepository<RecoverCaseEntity, Long> {
+public interface RecoverCaseRepository extends JpaRepository<RecoverCaseEntity, Integer> {
 }
